fix: remove old nvim configs. fix wezterm git plugins on work computer
Work computer uses a global git config to rewrite http git urls to ssh
git urls. In order to make it work correctly with wezterm that needs to
be undone when pulling wezterm plugins only. In order to do that an
overwrite gitconfig needs to be provided as
`~/dotfiles/wezterm/.config/wezterm/.gitconfig` and the global config
needs the following

```
[includeIf "gitdir:~/Library/Application Support/wezterm/plugins/"]
    path = ~/.config/wezterm/.gitconfig
[url "ssh://git@github.com/"]
    insteadOf = "https://github.com/"
```

Instead of just

```
[url "ssh://git@github.com/"]
    insteadOf = "https://github.com/"
```